reimpl: go back to storing coordinates in Section vs Grid, but with simple (if not super efficient) section-collision math.
Placeholders are now _derived_ instead of stored as a Section type,
by looking at every coordinate of the grid, if it falls within any of the
sections.

Note as with previous implementations, it doesn't handle well truncating
or removing sections when the grid is reduced in size. I think we could get awa
with truncating only; it's kind of nice to keep the sections "out of grid" so
if you expand the grid again they appear.